Extend HPA stats counters
There are few specific events we'd like to know more about to improve
purging strategy.

Empty pages are special case, we'd like to produce more of them and
purge them. So, `nempty_purges` counter was introduced to control how
good we at it.

We can think about purging strategy as cache eviction policy. In that
case we are interested how often we «evicted» memory and then requested
more of it, becase couldn't place following allocation anywhere. To
track that we need to track # of empty pages used (`nempty_used`)
and # of times we extracted fresh hugepage from central (`nextracted`).

There is a hypothesis kernel is not very good at assembling hugepages
back when we punched too much holes in them. For track how much we
are hugifying and dehugifying hugepages with holes `nhugifies_purged`
and `ndehugifies_purged` counters were introduced.